book dao 마이바티즈 구현

+++ 02_java/book/src/book/vo/Code.java
@@ -1,6 +1,6 @@
 package book.vo;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 /**
  * CODE 테이블의 컬럼과 동일한 vo 클래스
@@ -9,13 +9,13 @@ import java.math.BigInteger;
  */
 public class Code {
 	
-	private BigInteger code;
-	private BigInteger pCode;
+	private BigDecimal code;
+	private BigDecimal pCode;
 	private String codeNm;
 	private String codeVal;
 	private String useYn;
 	private String codeDesc;
-	private BigInteger sortOrder;
+	private BigDecimal sortOrder;
 	private String regId;
 	private String regDate;
 	private String modId;
@@ -25,19 +25,19 @@ public class Code {
 		super();
 	}
 
-	public BigInteger getCode() {
+	public BigDecimal getCode() {
 		return code;
 	}
 
-	public void setCode(BigInteger code) {
+	public void setCode(BigDecimal code) {
 		this.code = code;
 	}
 
-	public BigInteger getpCode() {
+	public BigDecimal getpCode() {
 		return pCode;
 	}
 
-	public void setpCode(BigInteger pCode) {
+	public void setpCode(BigDecimal pCode) {
 		this.pCode = pCode;
 	}
 
@@ -73,11 +73,11 @@ public class Code {
 		this.codeDesc = codeDesc;
 	}
 
-	public BigInteger getSortOrder() {
+	public BigDecimal getSortOrder() {
 		return sortOrder;
 	}
 
-	public void setSortOrder(BigInteger sortOrder) {
+	public void setSortOrder(BigDecimal sortOrder) {
 		this.sortOrder = sortOrder;
 	}
